+++ src/stories/Actions.stories.ts
@@ -5,7 +5,7 @@ import { withEngine } from './utils';
 import heartTexture from './assets/heart.png';
 
 export default {
-  title: 'Actions',
+  title: 'Actors/Actions',
   decorators: [withKnobs]
 };
 
@@ -33,6 +33,6 @@ export const fade: Story = withEngine(async (game) => {
 fade.story = {
   parameters: {
     componentSubtitle: 'Fade action',
-    notes: 'Use `Actor.fade()` to fade in or out the Actor over time'
+    docs: { storyDescription: 'Use `Actor.fade()` to fade in or out the Actor over time' }
   }
 };